logger/requirements.txt
flask==2.0.1
werkzeug==2.0.1
google-cloud-storage==2.9.0